# docs/api/public-api.openapi.yaml
openapi: 3.0.3
info:
  title: OpenGlass Public API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: |
    REST contract for the OpenGlass public layer (MVP scope).

    Source of truth for both `MockApiClient` and the Go backend — see
    `docs/dev/frontend-first-development.md`. No invented wire fields:
    additions require amending this file first.

    Conventions:
    - All endpoints are under `/api/v1`.
    - Invite issuance is an **ops-side operation** (server CLI / admin
      tooling on the instance host), deliberately not part of the public
      client API — there is no "create invite" endpoint here.
    - Auth uses `Authorization: Bearer <access_token>` (JWT, short-lived)
      plus a refresh token cookie/body flow.
    - Error envelope and codes: `docs/api/errors.md`.
    - Realtime events are out of scope here — see `docs/api/ws-events.md`.
servers:
  - url: /api/v1
...
paths:
...
  /contacts:
    get:
      tags: [contacts]
      summary: Contact list
      operationId: listContacts
      security: [{ bearerAuth: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [contacts]
                properties:
                  contacts:
                    type: array
                    items: { $ref: '#/components/schemas/Contact' }
    post:
      tags: [contacts]
      summary: Add contact (S7)
      operationId: addContact
      description: |
        Creates a one-directional contact entry. When both users have added
        each other the relation becomes `mutual` — required for 1-1 chat
        creation and private-session invites.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [contact]
                properties:
                  contact: { $ref: '#/components/schemas/Contact' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already in contacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /contacts/{userId}:
    delete:
      tags: [contacts]
      summary: Remove contact
      operationId: removeContact
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats:
    get:
      tags: [chats]
      summary: Chat list (S4)
      operationId: listChats
      description: Ordered by last activity; includes 1-1 chats and groups.
      security: [{ bearerAuth: [] }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [chats]
                properties:
                  chats:
                    type: array
                    items: { $ref: '#/components/schemas/ChatSummary' }
    post:
      tags: [chats]
      summary: Open or create a 1-1 chat with a mutual contact (S7 → S5)
      operationId: openDirectChat
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Existing chat returned
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '201':
          description: New chat created
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '403':
          description: Not mutual contacts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /chats/{chatId}:
    get:
      tags: [chats]
      summary: Chat details (S5 header / S7 actions)
      operationId: getChat
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [chat]
                properties:
                  chat: { $ref: '#/components/schemas/Chat' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatId}/pin:
    post:
      tags: [chats]
      summary: Pin/unpin chat in own chat list (S4 ordering)
      operationId: setChatPinned
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pinned]
              properties:
                pinned: { type: boolean }
      responses:
        '204':
          description: Updated
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /chats/{chatId}/messages:
    get:
      tags: [messages]
      summary: Message history, cursor-paginated (S5)
      operationId: listMessages
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
        - name: before
          in: query
          schema: { type: string }
          description: Opaque cursor from `nextCursor` of a previous page
            (older messages). Server implements it over the per-chat `seq`;
            clients must treat it as opaque.
        - name: after
          in: query
          schema: { type: string }
          description: Opaque cursor from `newerCursor` — paginates toward
            newer messages (forward scroll after an `around` jump).
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: q
          in: query
          schema: { type: string, minLength: 1 }
          description: Full-text search within this chat (S5b). When set,
            results are matching messages ordered by seq desc; `before`
            cursor paginates further matches.
        - name: pinned
          in: query
          schema: { type: boolean }
          description: When true, returns only pinned messages (S5b pinned view)
        - name: around
          in: query
          schema: { type: string, format: uuid }
          description: |
            Jump-to-message (search results, pinned messages, replies):
            returns a window of `limit` messages centered on this message's
            seq. Mutually exclusive with `q` and `pinned`.
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [messages, nextCursor, newerCursor]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor toward older messages; null at history start
                  newerCursor:
                    type: string
                    nullable: true
                    description: |
                      Cursor toward newer messages — only non-null after an
                      `around` jump, letting the client scroll forward from
                      the jumped position; pass it in the `after` param.
                      null when already at the latest.
    post:
      tags: [messages]
      summary: Send a message (S5)
      operationId: sendMessage
      description: |
        Server assigns `id`, `seq`, `sentAt`. Delivery to online members
        happens via `message.new` WS event — see `docs/api/ws-events.md`.
        Optimistic-UI/outbox semantics are defined by the contract: client
        sends `clientNonce`; server echoes it in the response and the WS
        event for dedup. The pair `(senderId, clientNonce)` is unique:
        retrying a send with the same nonce returns the already-created
        message with `200` (never a duplicate), so outbox flush is safe.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientNonce]
              properties:
                clientNonce:
                  type: string
                  description: Client-generated idempotency key (uuid)
                text:
                  type: string
                  maxLength: 8192
                  description: Required unless `attachmentIds` is non-empty
                attachmentIds:
                  type: array
                  items: { type: string, format: uuid }
                  maxItems: 10
                  description: |
                    Required unless `text` is non-empty. Each id must
                    reference an attachment uploaded to this chat
                    (`POST /chats/{chatId}/attachments`), else
                    `validation_failed`.
                replyToMessageId:
                  type: string
                  format: uuid
                  description: Must reference a message in this chat
      responses:
        '200':
          description: Idempotent replay — `clientNonce` was already used,
            returns the existing message
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /messages/{messageId}:
    patch:
      tags: [messages]
      summary: Edit own message
      operationId: editMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 8192 }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [messages]
      summary: Delete a message
      operationId: deleteMessage
      description: Own messages; in groups also users with `delete_messages` right.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /messages/{messageId}/pin:
    post:
      tags: [messages]
      summary: Pin/unpin a message (S5b pinned view)
      description: |
        Direct chats: either party. Groups: requires `pin_messages`
        right or owner.
      operationId: setMessagePinned
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/MessageId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pinned]
              properties:
                pinned: { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /chats/{chatId}/read:
    post:
      tags: [messages]
      summary: Mark chat read up to a message (read receipts)
      operationId: markRead
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upToSeq]
              properties:
                upToSeq: { type: integer }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '204': { description: Recorded }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ChatId:
      name: chatId
      in: path
      required: true
      schema: { type: string, format: uuid }
    MessageId:
      name: messageId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing/expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: "Token expired" } }
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    User:
      type: object
      required: [id, displayName, tag, createdAt]
      properties:
        id: { type: string, format: uuid }
        displayName:
          type: string
          description: Free-form, changeable, NOT unique
        tag:
          type: string
          description: Immutable unique handle `name#NNNN`; `#1` reserved for instance owner
          example: "jojou#1042"
        avatarUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    Contact:
      type: object
      required: [user, mutual, addedAt]
      properties:
        user: { $ref: '#/components/schemas/User' }
        mutual:
          type: boolean
          description: Both parties added each other
        addedAt: { type: string, format: date-time }
        verifiedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Last identity verification of this contact (US-2.5/2.6).
            Written by the private layer; always null until the private
            module ships. Drives the optional freshness reminder.
    Chat:
      type: object
      required: [id, type, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [direct, group] }
        title:
          type: string
          nullable: true
          description: Group title; null for direct chats
        peer: { $ref: '#/components/schemas/User' }
        members:
          type: array
          items: { $ref: '#/components/schemas/GroupMember' }
        createdAt: { type: string, format: date-time }
    ChatSummary:
      type: object
      required: [id, type, lastActivityAt, unreadCount]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [direct, group] }
        title: { type: string, nullable: true }
        peer: { $ref: '#/components/schemas/User' }
        lastMessage: { $ref: '#/components/schemas/Message' }
        lastActivityAt: { type: string, format: date-time }
        unreadCount:
          type: integer
          description: Per-user, synced across the user's devices
        memberCount:
          type: integer
          description: Group chats only — for the "N members" row in S4
        pinned:
          type: boolean
          description: User-side pin; pinned chats sort before unpinned,
            then by lastActivityAt desc
    MemberRights:
      type: object
      description: Party/raid model — owner grants specific rights per member
      properties:
        inviteMembers: { type: boolean, default: false }
        removeMembers: { type: boolean, default: false }
        editInfo: { type: boolean, default: false }
        pinMessages: { type: boolean, default: false }
        deleteMessages: { type: boolean, default: false }
    GroupMember:
      type: object
      required: [user, role, rights, joinedAt]
      properties:
        user: { $ref: '#/components/schemas/User' }
        role: { type: string, enum: [owner, member] }
        rights: { $ref: '#/components/schemas/MemberRights' }
        joinedAt: { type: string, format: date-time }
    Attachment:
      type: object
      required: [id, kind, mimeType, sizeBytes, url]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [photo, file] }
        mimeType: { type: string }
        fileName: { type: string }
        sizeBytes: { type: integer }
        url:
          type: string
          description: Authorized URL — requires the same bearer token,
            not a public link (server-trusted, not public internet)
        thumbnailUrl: { type: string, nullable: true }
    Message:
      type: object
      required: [id, chatId, seq, senderId, sentAt, clientNonce]
      properties:
        id: { type: string, format: uuid }
        chatId: { type: string, format: uuid }
        seq:
          type: integer
          description: Per-chat monotonic sequence — ordering + read cursors
        senderId: { type: string, format: uuid }
        text: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/Attachment' }
        replyToMessageId: { type: string, format: uuid, nullable: true }
        clientNonce:
          type: string
          description: Echoed for optimistic-UI dedup
        sentAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
        deletedAt: { type: string, format: date-time, nullable: true }
        pinned: { type: boolean }